split go cli command files
