+++ apps/cli/cmd/newsjack/config.go
@@ -1,0 +1,13 @@
+package main
+
+import ()
+
+const (
+	version                 = "0.2.0-go"
+	defaultRepo             = "elvisun/newsjack"
+	defaultRef              = "main"
+	defaultMinQueuePriority = 40.0
+	defaultMinMajorNews     = 0.55
+	medialystMCPURL         = "https://medialyst.ai/api/mcp"
+	envMedialystKey         = "MEDIALYST_API_KEY"
+)
