+++ apps/cli/cmd/newsjack/detector_test.go
@@ -1,0 +1,40 @@
+package main
+
+import (
+	"bytes"
+	"encoding/json"
+	"testing"
+)
+
+func TestDetectorMockRunAcceptsFlagsAfterQuery(t *testing.T) {
+	repo := repoRootForTest(t)
+	withTempEnv(t, map[string]string{
+		"HOME":          t.TempDir(),
+		"NEWSJACK_ROOT": repo,
+	}, func() {
+		var out, err bytes.Buffer
+		code := runCLI([]string{"detector", "run", "AI customer support", "--mock", "--include-all-scored", "--emit", "json"}, &out, &err)
+		if code != 0 {
+			t.Fatalf("detector code=%d stderr=%s", code, err.String())
+		}
+		var payload map[string]any
+		if json.Unmarshal(out.Bytes(), &payload) != nil {
+			t.Fatalf("invalid JSON: %s", out.String())
+		}
+		monitor := valueOrEmptyMap(payload["monitor"])
+		if monitor["mock"] != true {
+			t.Fatalf("mock=false payload=%s", out.String())
+		}
+		signals := signalSlice(payload["signals"])
+		if len(signals) != 2 {
+			t.Fatalf("signals=%d, want 2", len(signals))
+		}
+		if signals[0]["id"] != "e32ebc6ac34ee9d2" || signals[1]["id"] != "578832fabe7e6a64" {
+			t.Fatalf("unexpected signal ids: %#v %#v", signals[0]["id"], signals[1]["id"])
+		}
+		debug := valueOrEmptyMap(payload["debug"])
+		if dropped := anySlice(debug["dropped_signal_ids"]); len(dropped) != 0 {
+			t.Fatalf("dropped=%v, want empty", dropped)
+		}
+	})
+}
