+++ apps/cli/cmd/newsjack/flags.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"strings"
+)
+
+type stringList []string
+
+func (s *stringList) String() string { return strings.Join(*s, ",") }
+func (s *stringList) Set(v string) error {
+	*s = append(*s, v)
+	return nil
+}
